pdf_invoice/requirements.txt
Django
celery
django-celery-beat
django-celery-results
reportlab
Pillow
requests
flower
